fix: Put the correct bytes in for span and trace ids
The `LogRecord` expects the `trace_id` and `span_id` fields to be bytes.
By putting in the hex_string in these fields directly is downstream
providers will essentially encode16 the already encoded 16 IDs.

By first calling `decode16` first we are properly putting in the raw
bytes of the IDs into the protobuf before sending down to providerss.