Isolate HTTP request to own task
We have observed :reuse messages received because of the usage of Req/HTTP requests in the process.

Rather than handle those messages in the process arbitrarily it feels like it makes more sense to simply isolate the request itself.

I've set an arbitrary 20 second timeout on the task but I could also set it to infinity.