fix: Handle missing ETS table during shutdown (#24)
We have observed errors (see
[Sentry](https://electricsql-04.sentry.io/issues/41603549/?environment=production&project=4508410462404688&query=is%3Aunresolved&referrer=issue-stream))
trying to handle metrics during shutdown where they try to write to a
table that does not exist.

Even with our handler deregistration it can still run into races of
handlers running during the deregistration.

This approach ensures that if we see an error and the table does not
exist, we deregister the handler safely without errors logged.